+++ config/routes.rb
@@ -1,14 +1,14 @@
 # frozen_string_literal: true
 
-require 'plugin_routes'
-require 'camaleon_cms/engine'
+require "plugin_routes"
+require "camaleon_cms/engine"
 
 Rails.application.routes.draw do
-  scope PluginRoutes.system_info['relative_url_root'] do
+  scope PluginRoutes.system_info["relative_url_root"] do
     # Admin Panel
-    scope :admin, as: 'admin', path: PluginRoutes.system_info['admin_path_name'] do
-      namespace 'plugins' do
-        namespace 'camaleon_sitemap_customizer' do
+    scope :admin, as: "admin", path: PluginRoutes.system_info["admin_path_name"] do
+      namespace "plugins" do
+        namespace "camaleon_sitemap_customizer" do
           controller :admin do
             get :index
             get :settings