+++ lib/camaleon_sitemap_customizer.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require 'camaleon_sitemap_customizer/engine'
+require "camaleon_sitemap_customizer/engine"
 
 module CamaleonSitemapCustomizer
 end