+++ test/dummy/db/migrate/20180526080541_create_db_structure.cama_contact_form_engine.rb
@@ -2,7 +2,7 @@
 # This migration comes from cama_contact_form_engine (originally 20160517143441)
 class CreateDbStructure < PluginRoutes.migration_class
   def change
-    unless table_exists? 'plugins_contact_forms'
+    unless table_exists? "plugins_contact_forms"
       create_table :plugins_contact_forms do |t|
         t.integer :site_id, :count, :parent_id
         t.string :name, :slug
@@ -10,9 +10,9 @@ class CreateDbStructure < PluginRoutes.migration_class
         t.timestamps
       end
     end
-    if ActiveRecord::Base.connection.table_exists? 'cama_term_taxonomy'
+    if ActiveRecord::Base.connection.table_exists? "cama_term_taxonomy"
       CamaleonCms::Site.all.each do |s|
-        s.plugins.where(slug: 'contact_form').update_all(slug: 'cama_contact_form')
+        s.plugins.where(slug: "contact_form").update_all(slug: "cama_contact_form")
       end
     end
   end