+++ test/dummy/db/migrate/20180526080542_post_table_into_utf8.camaleon_cms_engine.rb
@@ -4,7 +4,7 @@ class PostTableIntoUtf8 < CamaManager.migration_class
     if table_exists? CamaleonCms::User.table_name
       add_column(CamaleonCms::User.table_name, :email, :string) unless column_exists?(CamaleonCms::User.table_name, :email)
       add_column(CamaleonCms::User.table_name, :username, :string) unless column_exists?(CamaleonCms::User.table_name, :username)
-      add_column(CamaleonCms::User.table_name, :role, :string, default: 'client', index: true) unless column_exists?(CamaleonCms::User.table_name, :role)
+      add_column(CamaleonCms::User.table_name, :role, :string, default: "client", index: true) unless column_exists?(CamaleonCms::User.table_name, :role)
       add_column(CamaleonCms::User.table_name, :parent_id, :integer) unless column_exists?(CamaleonCms::User.table_name, :parent_id)
       add_column(CamaleonCms::User.table_name, :site_id, :integer, index: true, default: -1) unless column_exists?(CamaleonCms::User.table_name, :site_id)
       add_column(CamaleonCms::User.table_name, :auth_token, :string) unless column_exists?(CamaleonCms::User.table_name, :auth_token)
@@ -23,7 +23,7 @@ class PostTableIntoUtf8 < CamaManager.migration_class
 
         # t.integer  "site_id",   default: -1, index: true
         t.timestamps null: false
-        t.belongs_to :site, index: true, default: -1#, foreign_key: true
+        t.belongs_to :site, index: true, default: -1 # , foreign_key: true
       end
     end
 
@@ -39,7 +39,7 @@ class PostTableIntoUtf8 < CamaManager.migration_class
       t.string   "status"
 
       t.timestamps null: false
-      t.belongs_to :user, index: true#, foreign_key: true
+      t.belongs_to :user, index: true # , foreign_key: true
     end
 
     create_table "#{PluginRoutes.static_system_info["db_prefix"]}posts" do |t|
@@ -56,7 +56,7 @@ class PostTableIntoUtf8 < CamaManager.migration_class
       t.string   "post_class", default: "Post", index: true
 
       t.timestamps null: false
-      t.belongs_to :user, index: true#, foreign_key: true
+      t.belongs_to :user, index: true # , foreign_key: true
     end
 
     create_table "#{PluginRoutes.static_system_info["db_prefix"]}term_relationships" do |t|
@@ -83,8 +83,8 @@ class PostTableIntoUtf8 < CamaManager.migration_class
       t.string   "agent"
       t.string   "typee"
       t.integer  "comment_parent", index: true
-      t.belongs_to :post, index: true#, foreign_key: true
-      t.belongs_to :user, index: true#, foreign_key: true
+      t.belongs_to :post, index: true # , foreign_key: true
+      t.belongs_to :user, index: true # , foreign_key: true
       t.timestamps null: false
     end
 
@@ -92,7 +92,7 @@ class PostTableIntoUtf8 < CamaManager.migration_class
       t.string  "object_class", index: true
       t.string  "name"
       t.string  "slug", index: true
-      t.integer  "objectid", index: true
+      t.integer "objectid", index: true
       t.integer "parent_id", index: true
       t.integer "field_order"
       t.integer "count", default: 0
@@ -118,8 +118,16 @@ class PostTableIntoUtf8 < CamaManager.migration_class
     end
 
     if ActiveRecord::Base.connection.adapter_name.downcase.include?("mysql")
-      ActiveRecord::Base.connection.execute "ALTER TABLE #{PluginRoutes.static_system_info["db_prefix"]}posts CONVERT TO CHARACTER SET utf8 COLLATE utf8_general_ci;" rescue nil
-      ActiveRecord::Base.connection.execute "ALTER TABLE #{PluginRoutes.static_system_info["db_prefix"]}custom_fields_relationships CONVERT TO CHARACTER SET utf8 COLLATE utf8_general_ci;" rescue nil
+      begin
+        ActiveRecord::Base.connection.execute "ALTER TABLE #{PluginRoutes.static_system_info["db_prefix"]}posts CONVERT TO CHARACTER SET utf8 COLLATE utf8_general_ci;"
+      rescue
+        nil
+      end
+      begin
+        ActiveRecord::Base.connection.execute "ALTER TABLE #{PluginRoutes.static_system_info["db_prefix"]}custom_fields_relationships CONVERT TO CHARACTER SET utf8 COLLATE utf8_general_ci;"
+      rescue
+        nil
+      end
     end
   end
 end