+++ test/dummy/db/migrate/20180526080547_improve_menus_structure.camaleon_cms_engine.rb
@@ -2,7 +2,7 @@
 class ImproveMenusStructure < CamaManager.migration_class
   def change
     CamaleonCms::NavMenuItem.all.each do |menu|
-      menu.update_columns({description: menu.get_option('object_id'), slug: menu.get_option('type')})
+      menu.update_columns({description: menu.get_option("object_id"), slug: menu.get_option("type")})
     end
   end
 end