+++ test/dummy/db/migrate/20180526080549_install_migrated_seo_plugin.camaleon_cms_engine.rb
@@ -3,7 +3,7 @@ class InstallMigratedSeoPlugin < CamaManager.migration_class
   # install seo plugin without calling hooks (seo logic moved to separated plugin)
   def change
     CamaleonCms::Site.all.each do |s|
-      s.plugins.where(slug: 'cama_meta_tag').first_or_create!(term_group: 1)
+      s.plugins.where(slug: "cama_meta_tag").first_or_create!(term_group: 1)
     end
   end
 end