+++ test/dummy/db/migrate/20180526080551_create_media.camaleon_cms_engine.rb
@@ -7,7 +7,7 @@ class CreateMedia < CamaManager.migration_class
       t.boolean :is_folder, index: true, default: false
       t.string :folder_path, index: true
       t.string :file_size
-      t.string :dimension, default: ''
+      t.string :dimension, default: ""
       t.string :file_type
       t.string :url
       t.string :thumb