+++ test/dummy/db/schema.rb
@@ -11,7 +11,6 @@
 # It's strongly recommended that you check this file into your version control system.
 
 ActiveRecord::Schema.define(version: 2018_05_26_080551) do
-
   create_table "active_storage_attachments", force: :cascade do |t|
     t.string "name", null: false
     t.string "record_type", null: false
@@ -203,5 +202,4 @@ ActiveRecord::Schema.define(version: 2018_05_26_080551) do
     t.datetime "created_at"
     t.datetime "updated_at"
   end
-
 end