+++ test/test_helper.rb
@@ -1,10 +1,10 @@
 # frozen_string_literal: true
 
-require File.expand_path('../test/dummy/config/environment.rb', __dir__)
-ActiveRecord::Migrator.migrations_paths = [File.expand_path('../test/dummy/db/migrate', __dir__)]
-require 'rails/test_help'
-require 'capybara/rails'
-require 'capybara/minitest'
+require File.expand_path("../test/dummy/config/environment.rb", __dir__)
+ActiveRecord::Migrator.migrations_paths = [File.expand_path("../test/dummy/db/migrate", __dir__)]
+require "rails/test_help"
+require "capybara/rails"
+require "capybara/minitest"
 
 Capybara.server_port = 3000
 
@@ -14,8 +14,8 @@ Minitest.backtrace_filter = Minitest::BacktraceFilter.new
 
 # Load fixtures from the engine
 if ActiveSupport::TestCase.respond_to?(:fixture_path=)
-  ActiveSupport::TestCase.fixture_path = File.expand_path('fixtures', __dir__)
+  ActiveSupport::TestCase.fixture_path = File.expand_path("fixtures", __dir__)
   ActionDispatch::IntegrationTest.fixture_path = ActiveSupport::TestCase.fixture_path
-  ActiveSupport::TestCase.file_fixture_path = ActiveSupport::TestCase.fixture_path + '/files'
+  ActiveSupport::TestCase.file_fixture_path = ActiveSupport::TestCase.fixture_path + "/files"
   ActiveSupport::TestCase.fixtures :all
 end